Consolidate App and Command structs
