+++ examples/nested_multiple_app.rs
@@ -1,4 +1,4 @@
-use seahorse::{error::FlagError, App, Command, Context, Flag, FlagType};
+use seahorse::{error::FlagError, App, Context, Flag, FlagType};
 use std::env;
 use std::error::Error;
 
@@ -58,8 +58,8 @@ fn hello_action(c: &Context) -> Result<(), Box<dyn Error>> {
     Ok(())
 }
 
-fn hello_command() -> Command {
-    Command::new("hello")
+fn hello_command() -> App {
+    App::new("hello")
         .description("hello command")
         .usage("multiple_app hello(he, h) [name]")
         .alias("h")
@@ -91,15 +91,15 @@ fn add_action(c: &Context) -> Result<(), Box<dyn Error>> {
     Ok(())
 }
 
-fn add_command() -> Command {
-    Command::new("add")
+fn add_command() -> App {
+    App::new("add")
         .description("add command")
         .usage("multiple_app add [num...]")
         .action(add_action)
 }
 
-fn world_command() -> Command {
-    Command::new("world")
+fn world_command() -> App {
+    App::new("world")
         .description("hello world command")
         .usage("nested_multiple_app hello(he, h) world(w)")
         .alias("w")
