+++ src/lib.rs
@@ -1,6 +1,5 @@
 mod action;
 mod app;
-mod command;
 mod context;
 pub mod error;
 mod flag;
@@ -9,7 +8,6 @@ mod utils;
 
 pub use action::Action;
 pub use app::App;
-pub use command::Command;
 pub use context::Context;
 pub use flag::{Flag, FlagType, FlagValue};
 use help::Help;
